remove unused code and packages
