+++ crates/client/state-sync/src/tests/mod.rs
@@ -1,5 +1,6 @@
 //! Madara client testing utilities.
 pub mod constants;
+pub mod ethereum;
 pub mod helpers;
 pub mod parser;
 pub mod sync;
